SYV682: Update Driver for B revision silicon
They SYV682B made some meaningful changes to FRS. Update to remove the
workaround put in place for the SYV682A and replace with the FRS logic
for the SYV682B. The RVS alert can now be masked, so remove the message
for an RVS event and repurpose the flag.

There are some other changes, but none require driver changes.

BUG=b:148144711
TEST=Enable FRS in board config, check that FRS works with with a 5V and
20V supply on a fast and slow-discharge FRS device.
BRANCH=none

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: I78e1c7d5eca593661a7d67346c83775ca990e38d