sc7180: Notify SUSPEND hooks in S0S3 if shutdown from on
This CL pairs with the previous CL:2391202, which notifies the RESUME
hook in S3S0 if boot from off. This CL does the counter part, which
notifies the SUSPEND hooks in S0S3 if shutdown from on.

BRANCH=None
BUG=b:148149387, b:170604357
TEST=Verified the SUSPEND hooks are called when shutdown from S0.
For example, all the PWM channels are disabled after power off AP.

Change-Id: Ib420d5d0e560f30f9c03b5e31788a145279a9c75
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2468476
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>