TCPMv2: Update mux state if Data role swap times out or NAKs
On receiving an ACCEPT for data role swap, the mux state is updated
according to the port partner's USB capability and port's current data
role.

But, on receiving a REJECT or timeout, even if the data role remains
unchanged, we still need to update the mux with port partner's USB
capability.

BUG=b:169935734, b:167595673
BRANCH=None
TEST=Able to charge the USB4 DB and able to update mux capabilities
on receiving a timeout or NAK on Data role swap

Signed-off-by: Ayushee <ayushee.shah@intel.com>
Change-Id: I80c9e6bcf82ffbf15a136cbf72b24fd37eb22cc6