ln9310: Enable LB_MIN_FREQ_EN
In order to mitigate the risk of overvolting something that uses
very low power, we want to make sure our switching frequency is
not too slow.

Also, we are choosing to use VIN for VDR in case EXT_5V is not
available when the chip is trying to startup.

TODO: This patch might have some power efficiency implications on
lower current loads, but until we figure out a more ideal fix
this is a good temporary mitigation.

BRANCH=master
BUG=b:170591575
TEST=With a non ideal switchcap, and no load (eg: on limozeen keep
qsip pmic off), touch with a finger the BST1A capacitor (eg C67)
After this patch nothing bad will happen, without this patch the
QSIP might get overvolted and permanently killed.

Signed-off-by: Alexandru M Stan <amstan@chromium.org>
Change-Id: Ie230eb8fe87db35854b727585a667cbfdb30e46c