SM5803: Clear out bad state in OTG disable
When sourcing fails to a port partner, the discharge status should be
cleared before clearing out the FLOW1 register.  Since the source bit
may be unreliable after sourcing fails, use this status as an indication
for whether it's safe to clear out the mode from FLOW1.

BRANCH=None
BUG=b:169966133,b:170517117
TEST=on drawlat, attach bad dongle and observe the port works with a
good dongle after the bad one is unplugged

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I128e292b3f2bfe5362e51b6967f62e4191809149