servo_updater: Add timeout on the pty flush
If the servo is in a loop where it sends output on a continuous basis,
the flush operation needs to eventually timeout. If the servo sends
commands out faster than the timeout provided to the expects() call,
it would never advance. This blocked the servo_updater from advancing.

BUG=b:162628581
TEST=Connect Servo_v4p1 to DUT validate servo_updater works
TEST=Modify Servo firmware to generate many console messages
    in a loop. Validate that the flush operation succeeds and
    servo updater still works.

Change-Id: I8c0e65a0733b4d4ee25a9f9c866aa2b863cb0e5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2487256
Reviewed-by: Matthew Blecker <matthewb@chromium.org>
Tested-by: Brian Nemec <bnemec@chromium.org>
Commit-Queue: Brian Nemec <bnemec@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>