SM5803: Only set default input current when battery is present
During init time, only limit to default input current if the battery
is physically present.  Otherwise, we may risk browning out the board
if the AP is on and the EC is sysjumping (ex. during EC flash).

BRANCH=None
BUG=b:168169691
TEST=on drawlat, unplug battery and verify running a sysjump with the AP
on doesn't brown out during init (instead, it browns out later when a
hard reset is sent)

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I700aabd51b45e2aff459f787c13978c6212026f5