Volteer: Add EC console command to get VCCIN power
VCCIN is the core regulator on TGL, but it can't be read via sense
resistors because of its high current.

This adds a command on the ec console 'vccin_pwr' which will give the
average power and voltage of vccin over 1 second. this will fail if
the AP is off.

BUG=none
TEST=Test if we can read VCCIN power on Volteer
BRANCH=none

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: I0fe94adccac9f2a2be625f8479d4ada8cfb9c016