ps8xxx: workaround SNK device not detected
The PS8xxx fails to start DRP toggling when a SNK device is connected
while the PD port is disabled due to S0ix, S3, or G3 states.

For the PS8xxx, send the Looking4Connection command before writing to
the ROLE_CONTROL register to set the pulls.

BUG=b:169632095
BRANCH=none
TEST=make buildall
TEST=Connect a SNK device while the AP is S0ix or G3 to a Volteer board.
On wakeup, observe DRP toggling starts and the device is detected.

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: Ica9bc50cf63f7c98473133944030965a4e511851