TCPMv2: Disallow hard resets without a battery
When a chromebook is booted without a battery and reaches the point
where it would normally hard reset a connection, disable the PE for the
port instead.  This may result in a strange state for the board (ex.
charge manager reporting the active port is a type-c supplier with
15V), but will not brown out the board during important factory
sequences.

Note that chromeboxes should brown out in this condition, as it would be
undesirable to leave a connection up with only partial functionality on
a customer system.

BRANCH=None
BUG=b:168169691
TEST=on drawlat, boot without a battery and run "flashrom -p ec"
from the OS successfully

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I0d27348532920a0c7e36c85f1fba24495ba7d648