quiche: Add support for C2 usbc port
The 3rd usbc port (C2) is type-c only port which is managed directly
by the PS8803 TCPC, but C2 has a PPC and the EC needs to control VBUS
on/off via the PPC.

This CL adds an interrupt handler to track the signal that indicates
when the TCPC is in Attached.SRC or Unattached.SRC. In addition, the
PPC for C2 must be initialized at the board level as this port is not
managed by the TCPMv2 stack.

BUG=b:171915198
BRANCH=None
TEST=Used C -> A adapter, verifed that when connected C2 is in
Attached.src and removed in Unattached.SRC. Validated bus signals with
TotalPhase trace. Also checked have a device connected to this port
enumerates at the host.

> [28.613330 C2: State = Unattached.SRC]
[36.500398 C2: State = Attached.SRC ]

[135240.133949] usb 1-2.2.1: new full-speed USB device number 87 ...
[135240.229223] usb 1-2.2.1: New USB device found, idVendor=046d, ...
[135240.229246] usb 1-2.2.1: New USB device strings: Mfr=1, ...
[135240.229257] usb 1-2.2.1: Product: BOOM 3
[135240.229267] usb 1-2.2.1: Manufacturer: Logitech
[135240.229276] usb 1-2.2.1: SerialNumber: 000001

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Ie2583ae9271333e2ecd9561eed60fab6cdb4fda1