OCPC: Average and filter resistance calculations
It's been seen that the resistance calculations can vary quite a bit
during the charging cycle.  In order to reduce this variance, this CL
has the OCPC module take the running mean and use that at the combined
resistance value.  Additionally, values that are outside of a 6
standard deviation window are discarded.

BUG=b:170974274
BRANCH=None
TEST=Build and flash waddledee, enable OCPC debug logging and plot
combined resistance over time as DUT charges in suspend.  Verify that
combined resistance does not vary significantly.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I2ccb9ccbc0b98455e78249d47f8692828f1a15e9