tcpmv2: ifdef cleanup with IS_ENABLED
Remove as many #if preprocessor statements as possible
without allowing the code to grow when it is compiled.

Verified against the sizes built on current ToT and
this CL on that ToT

~/trunk/src/platform/ec $ make newsizes
Compared 699 of 699 files.
File sizes are unchanged.

BUG=b:172208011
BRANCH=none
TEST=verify TCPMv2 functionality

Signed-off-by: Denis Brockus <dbrockus@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2515279
Tested-by: Denis Brockus <dbrockus@chromium.org>
Auto-Submit: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Jett Rink <jettrink@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Change-Id: Idb9a6bca9aa2710bb09d8ef9d5a2b874d5e9c6e7