Trogdor: Request 5V when AP is off and battery is full
The charger is not efficient on buck mode when the operating current
is low. It happens when AP is off and battery is full.

Request the external adapter to a lower voltage, 5V. The charger will
switch to boost mode that has a better efficiency.

BRANCH=None
BUG=b:170794007
TEST=Charged the battery to full. Checked it requesting 20V when AP is
ON, and requesting 5V when AP is OFF.
TEST=Discharged the battery to not full. Checked it requesting 20V no
matter AP is ON or OFF.

Change-Id: Ifa333fccb40e7ebb2461c396be392c4ccc2d4320
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2518326
Reviewed-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>