honeybuns: usbc default Rp and baseboard hook init timing
This CL changes the hook init timing for baseboard_init which needs
to be run prior to any usbc/usb-pd initialization. In addition the
config option CONFIG_USB_PD_PULLUP is changed from RP_1A5 to RP_3A0 to
match what is advertised in the SRC CAP message.

BUG=None
BRANCH=None
TEST=Verfied TD.4.9.2 USB Type C Current Advertisement Test passes

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I226e983733850f1b21c25175826e762295609069