TCPMv2: Add custom protocol error handling for FRS
For pe_report_error(), there are several cases which are handled
differently than the default (soft reset), which are described in
PD3.0 section 8.3.3.4.1.1. One of these is a protocol error during an
FRS Send_Swap, which should be a transition to ErrorRecovery instead of
Soft Reset (section 8.3.3.18.6.2). A normal PRS Send_Swap follows the
default behavior.

Add this exception to the default handling of protocol errors. Without
this, the TCPM/TCPC will get into an incorrect state which isn't
recoverable without a manual reset.

BUG=b:171461736
TEST=make buildall
TEST=On volteer, with FRS enabled, check that removing a device while
primed for FRS results in a transtion to ErrorRecovery instead of
Soft Reset.
BRANCH=None

Change-Id: I350aa39d6f41e7a1d304389fd8bf2e7693cd912c
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>