sm5803: Add hardware charging ramping
This patch add support hardware charging ramp. The hw ramp is
implemented by enabling DPM (Dynamic Power Management). Once a
DPM Voltage loop regulation is set and the DPM loop is enabled,
the DPM guarantees that the Vbus will not go below the threshold
voltage limiting the power that can be delivered to the load.

BUG=b:172173517
BRANCH=none
TEST=On drawcia, i2ctrace to make sure DPM is enabled with 5V source.

Signed-off-by: Devin Lu <Devin.Lu@quantatw.com>
Change-Id: I0622da35bf57f367b21e97c5e39cb157ef911521