zephyr: Enable the '8042' command
This is used to provide information about the keyboard operation.
Enable this command and its subcommands.

Note that in ECOS these subcommands are also top-level commands, but
this is not the case in Zephyr, since it seems unnecessary.

BUG=b:167405015
BRANCH=none
TEST=make BOARD=volteer
zmake configure .../zephyr-chrome/projects/experimental/volteer \
-B /tmp/z/cos
zmake build /tmp/z/cos
Run on volteer and try out '8042 kbd', etc.

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ia11092af350247ac98681485a9ddd309c7192272