raa489000: Ignore BATGONE on aux chargers
On OCPC systems, the BATGONE pin is left unconnected as it's not
needed.  However, the default setting of the RAA489000 looks at the
BATGONE pin in order utilize LEARN mode.  This commit modifies the
initialization to ignore the BATGONE pin on auxiliary chargers in an
OCPC system.

BUG=b:171457154
BRANCH=None
TEST=Build and flash magolor, verify that `chgstate discharge on`
results in 0 amperes being consumed from the connected AC adapter.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Iefec97e2dfcf0852a02c7acd05257014e93e9f57