servo_v4p1: Increase RW flash size
Size of RO section was reduced from 92K to 90K in favor of space for RW
part. This value was chosen experimentally, so that new changes to the
i2c_master fits in.

This value may be adjusted further, however most of the functionalities
on servo_v4p1 are placed in RO section.

BUG:b:150323106
BRANCH:master
TEST:buildall.

Signed-off-by: Jan Dabros <jsd@semihalf.com>
Change-Id: Ia80d5f3799743aaa65a01223f1864706a10d3c68