power: allow eSPI virtual wires without eSPI host commands for zephyr
For Zephyr OS, we want to get eSPI up and going for power sequencing
without having to bring in all of the host command dependencies.
Right now the power sequencing code assumes that if we are using eSPI
for host commands, that means we might have eSPI virtual wires too.
Instead, use the separate option CONFIG_HOST_ESPI_VW_POWER_SIGNAL,
which is what we actually want, and allow that to be defined without
CONFIG_HOSTCMD_ESPI.

BUG=b:171312361
BRANCH=none
TEST=buildall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I2f56ef3ab9cc566f5e0e3926fea96484daa93236
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2548302
Reviewed-by: Yuval Peress <peress@chromium.org>
Signed-off-by: Wai-Hong Tam <waihong@google.com>