zephyr/test/base32: change test to depend on less
As we add more shim code and configs, we don't want to remove the
functionality in all of the tests. Convert this test to only
include the files it needs to link to perform the shim test.

BRANCH=none
BUG=b:172512307
TEST=base32 test builds and passes

Signed-off-by: Jett Rink <jettrink@chromium.org>
Change-Id: I4a7232c98397600d578e18b67935924b088f377b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2532687
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
(cherry picked from commit a9691611c515d7ef57aa69d738909d607461b7b1)