TCPMv2: Always probe sink caps
This commit adds the ability to always probe the partner for sink
capabilities, and re-names the state to reflect that sources may use it
now too.  This also fixes some issues in the state to get sink
capabilities, including correcting the message detection and soft
resetting with the SOP* of the unexpected message.

This also updates the unit tests to expect a Get_Sink_Caps message as a
part of typical startup.

BRANCH=None
BUG=b:168862110,b:160009733
TEST=on drawcia, observe sink caps are requested from source and sink
partners

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I5da3adc17ccec71f6cad9e8f152b9570ec99691d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2540389
Reviewed-by: Keith Short <keithshort@chromium.org>
Signed-off-by: Wai-Hong Tam <waihong@google.com>