TCPMv2: Always probe sink caps
This commit adds the ability to always probe the partner for sink
capabilities, and re-names the state to reflect that sources may use it
now too.  This also fixes some issues in the state to get sink
capabilities, including correcting the message detection and soft
resetting with the SOP* of the unexpected message.

BRANCH=None
BUG=b:168862110,b:160009733
TEST=on drawcia, observe sink caps are requested from source and sink
partners

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I5da3adc17ccec71f6cad9e8f152b9570ec99691d