firmware_builder: add hardware test builds
We were not building the hardware test images as part for the test step.
This replicate functionality in the chromeos-ec ebuid (via cros-ec
eclass). I put a note in that ebuild to update this script if there are
any changes in the future.

BRANCH=none
BUG=chromium:1147953
TEST=verify that firmware_builder.py --metrics /tmp/m test failed the
same was the cq-orchestrator failed

Signed-off-by: Jett Rink <jettrink@chromium.org>
Change-Id: I20937381da41d933261926d02d45b80d3920443b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2535310
Reviewed-by: LaMont Jones <lamontjones@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>