sc7180: Fix forever loop if not enough power to boot AP
When the power supply is not enough (<10W adapter and no battery)
to boot AP, should clear the boot_from_off flag and go back to S5;
otherwise, the boot_from_off flag will trigger the power-on again.
Also do the same on power_on() sequence failed.

Increase the delay of checking power, such that the messages are
not too noisy.

BRANCH=None
BUG=b:167155164
TEST=Unplugged the battery, plugged a <10W adapter, checked the
"Not enough power to boot (-1 %, 7500 mW)" message just showed 10x.

Change-Id: Id7c5e66dab62bc71d3e2e00b3be172e13a146ed5
Signed-off-by: Wai-Hong Tam <waihong@google.com>