Trogdor: Ignore HPD Low if the DP path can't be muxed
Only one DP path is allowed in the Trogdor design. If the DP path
is occupied by another port, we ignored the HPD High but we forgot
to ignore the HPD Low. We should ignore both HPD High and Low.

When exit the DP alt-mode, also check if it is the correct path
before clean up.

BRANCH=None
BUG=b:173213397
TEST=Plugged the 1st monitor to a port. Plugging the 2nd monitor to
another port didn't affect the 1st monitor. Unplugging the 2nd monitor
didn't affect the 1st monitor.

Change-Id: Ice2ebc512239f062d22fa69befcaa39b9090c7db
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2537157
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>