genvif: USB-IF cleanup and --no-config
A couple of the lines would not compare correctly
because I wanted hex string values to have leading
0x and USB-IF does not.  USB-IF also does not
always set "YES" and "NO" for boolean tags, so I
added an override to make this happen on my output.
They also did not add the string value for XID_SOP
so I removed it from mine as well.

genvif has a new option of creating an output
file based solely on the overrides and not the
configs. This will make comparisons from other
generators a little easier.

USAGE: genvif -b|--board <board name>
              -o|--out <out directory>
              [-n|--no-config]
              [-v|--over <override XML file>]

BUG=b:173219559
BRANCH=none
TEST=verify XML output

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I4af5662321e29ed806dea0a2420fd8f4627f8e0f