motion: Make CONFIG_MOTION_SENSE_..._DELAY_US #undef when 0
Normally, we should change the sensor rate while is suspend/resume.
However, on some board, the power line to sensors are switched off
asynchronously, so we need a delay before acting on the sensors.

When we set the delay to 0, we were still processing the sensor rate in
a different state, where the sensor state has already been updated, so
we are not allowed to change the ODR anymore.

BUG=b:170703322
BRANCH=kukui
TEST=compile

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I0642ac6cda3e87dd11cc76c83581a21605b28e87