Trogdor: Add the interrupt for USB-A overcurrent
The USB-A VBUS is through a load switch. The FAULT_L pin of
the load switch requires an external pull-up. Apply the pull-up
on EC side.

Add the interrupt of the FAULT_L pin. Route to the function
board_overcurrent_event(), which currently prints an error
message.

BRANCH=Trogdor
BUG=b:173154219
TEST=Plugged a USB load >1.5A to the USB-A port and checked
the error message was printed:
[73.738372 p2: overcurrent!]

Change-Id: I5f5f82de942eaf26a33051ad17f14b90d8024b89
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2547190
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Signed-off-by: Keith Short <keithshort@chromium.org>