berknip: delay RTD2141B power-on for correct HPD
A firmware change to the MST hub has changed the behavior of the HPD
line to support in-system programming of hub firmware such that the HPD
line idles high instead of low when no sink is connected. That change
means the AP may miss HPD on resuming from suspend, requiring an
external display to be re-plugged if it was connected while the system
was suspended.

This change introduces a delay between system resume and enabling power
to the MST hub to help ensure the AP is awake before the hub may
generate HPD interrupt pulses.

BUG=b:171678892
TEST=still builds
BRANCH=zork

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Idadc6b8206f1c9c58fa398304932a6c1803dd833