Use host's pkg-config for gen_touchpad_hash
This is a helper running on the host, so should use host tooling.
Changing this avoids a build failure on cheza where chromeos-ec the
would fail to build before the openssl package (which is supposed to
work given that chromeos-ec doesn't list openssl in its dependencies,
and this pkg-config invocation is not a reason to start doing so).

BUG=None
BRANCH=None
TEST=emerge-cheza -v1 chromeos-ec

Signed-off-by: Mattias Nissler <mnissler@chromium.org>
Change-Id: I45161285db17a56b29a36549d3b113f42dff9362