bloonchipper: enable low power
Enable STOP mode in idle when the PLL is off.
On systems with serial host interface, disable the PLL when the system
is in S3.
On systems with SPI host interface, disable the PLL when we are not
computing fingerprints or RW hash.

Use only the SLP_S3_L signal to detect the AP state as SLP_S0_L behaves
differently on Zork and Hatch.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:130561737
BRANCH=fpmcu-bloonchipper
TEST=manual, on bloonchipper, check we can still capture fingerprint.
TEST=manual, read the MCU power consumption: pp3300_dx_mcu_mw is 2.367 mW.
TEST=manual, verify USART and SPI host interfaces.
TEST=manual, verify fingerprint recognition latency.

Change-Id: Ia6779238d65022156a53ed4d040620c8c8cfc85e