PD: Fix passing information about polarity when DTS is connected
Commit 7dec638eb577aaa3a00d0551d73c276b94ebacb2 introduced two polarity
modes POLARITY_CC1_DTS, POLARITY_CC2_DTS in enum tcpc_cc_polarity, but
in many places there was an assumption that value other than 0 means
that cable is inverted, the most notable example is usb_mux_set(). As a
result kernel sometimes was not reporting SuperSpeed depending on if
cable was inverted or not.

This patch adds mapping from polarity with DTS to polarity without DTS
where necessary.

BUG=b:162254118
BRANCH=none
TEST=Connect ServoV4 to eve and run servod. Make sure that USB-C muxer
     is connects USB3.0 lines (servod should set it).
     Flash EC ToT on eve. Boot ChromeOS and go to Developer Console.
     Run 'dmesg -w', check if device (eg. ethernet adapter) is attached as
     SuperSpeed device. Unplug cable, invert and plug again. Kernel
     should report that device is attached as SuperSpeed.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I354ef7047240cc8b5db01936b3780fae7387edb5