TCPMv2: AP resume should consider Src->Snk swap
TCPMv1 used PD_FLAGS_CHECK_PR_ROLE and
pd_check_pr_role() to accomplish this. This was
missing from TCPMv2.

In the TCPMv2 implementation, policy
pd_can_source_from_device was added to put the
logic in one place to detect if the PDO for a
partner device can and should be the source

Using this policy the AP resume code will check
to see if an already attached partner that we
are sinking from should be sinking from us instead.

BUG=b:169299049
BRANCH=zork
TEST=verify DRP phone will swap back to SNK on AP resume

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I8e584446445c8ee2e1c91973a58a04405cdf9e1b