Madoo: Improve shared interrupt line behavior
When interrupts are interleaved from the TCPC and charger, we can get
into a state where the interrupt line for the board ends up stuck low.
Improve on this situation by setting a deferred function to check on the
interrupt line after we get a signal in, and notify chips on that line
there may be more to process.

BRANCH=None
BUG=b:173345295
TEST=on madoo, plug in powered PD hub to C1, generating interrupts
from both the charger and TCPC and verify C1 interrupt line goes high
once connection is ready

Signed-off-by: Ko_Ko <Ko_Ko@compal.corp-partner.google.com>
Change-Id: I3d29df28f94cb6c352adf05b9204ab4ff52a9726