USB4: Update USB4 active cable flow
For the sideband channel of a router to operate as a USB4
sideband channel, the link over an active cable should be
unidirectional.

So, this CL adds a function to return the LSRX link-training
directionality for USB4 mode with active cable to unidirectional
and as per Discover Mode SOP' bit23 for Thunderbolt mode and USB4
mode with passive cable.

This CL also corrects setting the pd control flags needed for the
usb_pd_control host command.

BUG=b:174167882
BRANCH=None
TEST=Able to set the link-training to unidirectional for USB4 with
     active cable.

Signed-off-by: Ayushee Shah <ayushee.shah@intel.com>
Change-Id: Idd451b8e5f90643308a38879a20c514a97ce1dc7