WIP: Experimental change for BBR mux sequencing
Before configuring burnside bridge, make sure to reset it. Then wait
100ms before configuring TCSS. In Voxel, rearrange the mux orders so
that the retimer is configured before the TCSS mux in the AP. Fix an
issue in the BBR retimer that the VID was wrong, add retries to the BBR
retimer writes.

BUG=b:173725284
TEST=On Voxel, check that 100 disconnect/reconnects of Gatkex creek
always enumerate as USB4g3, never USB4g2.
BRANCH=none

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: I8c887839386723b3b36f384df7692ee3d06ed8d1