ocpc: Center hysteresis around target
We add a bit of hysteresis when determining our error for the OCPC PID
loop equivalent to charger IC's "step" in regulating current.  We had
set the band above and below the target for the full step.  In order
to reduce overshoot, this commit centers the band around the current
target.

BUG=b:174167890
BRANCH=dedede
TEST=Build and flash drawlat; verify that steady state now settles
around a shorter band and overshoot is reduced.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I8f8c7305c3973fb5cedecdd14e5466f11a0e85a7