zephyr: fix build for keyboard shim
See platform/ec/common/build.mk for reference. The use of
keyboard_scan.c should only be allowed when HAS_TASK_KEYSCAN
is defined. As such, we should always set HAS_TASK_KEYSCAN to
1 if CONFIG_PLATFORM_EC_KEYBOARD is defined. Further,
CROS_EC_TASK_LIST was defined per project and it doesn't need
to be, we can instead define it once for all boards depending
on which tasks are set to 1 in shimmed_tasks.h

Note that shimmed_tasks.h can still be used in tests.

BRANCH=none
BUG=none
TEST=zmake testall (this is technically still failing because
     of a missing CC_EVENTS).

Cq-Depend: chromium:2566421
Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: Id0ed49dd49e3c4eb2ac23184cf943c91dcd261eb