OCPC: Use `charger_is_icl_reached()` if possible
Some charger ICs have the capability to report if the input current
has been reached.  Thus far, it seems to be more precise compared to
the values returned via the charger ICs ADCs.  This commit has the
OCPC module use that functionality if available.  If the feature is
not available, the fallback will be what exists today, which is simply
seeing if the instnantaneous input current limit is 95%+ of the set
input current limit.

BUG=b:174167890
BRANCH=dedede
TEST=Build and flash drawlat.  Charge from sub board, verify that
input current is reached prior to that reported from the onboard ADCs.

Change-Id: I3dfb579fd4617b13dfdbed3cb9c7cc2c2f70b93a
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2568565
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>