WIP: TCPMv1: TCPMv2: Remove PD message queue
It takes approx. 3ms to dequeue a pd message into the local
buffer. This CL removes the queue and copies the pd message
from the TCPC directly to the local buffer. This reduces
the time to approx. 2ms.

This is a fix is for spammy docks such as the HP G5
that sometime sends Attention message at 3ms intervals,
resulting in the queue overflowing.

The debug statements used to measure time intervals are included
in this CL but will be removed if +2d.

To enable this patch, define CONFIG_USB_PD_MSG_DIRECT_COPY

Works with TCPMv1 and TCPMv2

BRANCH=none
BUG=:b:155517411
TEST=make -j buildall

Tested on Hatch, Atlas, and Ezkinil. Verified that HP G5 Dock
extended display to WIMAXIT monitor. Port 1 worked on all
device but Port 0 on Hatch and Ezkinil had problems that are
not related to this CL.

Unit tests need implementing

Change-Id: I47a7c0a319803c6e99d006412f84c65e3749409e