asurada: cache board id on init
The line board_get_version() inside board_hibernate_late() fails if it
is the first board_get_version() call, because ADC (or some function
related to it) is not usable at that time.

To fix this, cache board id to ensure board_hibernate_late() never need to
access ADC.

Also replace #include <...> to #include "..." for consistency.

BUG=b:163963220
TEST=hibernate successful in G3
BRANCH=none

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: I0c277cccd081906d6f87793e8e4f1980a4e55c3a