zephyr: Sort the options a bit
We have quite a few options already and plan to add more. Sort them to
make it easier to see what is there.

The I2C enum isn't really connected with the basic CONFIG munging done
in config_chip.h so move it to i2c.h instead.

Fix the 'keybaord' typo while we are here.

BRANCH=none
BUG=none
TEST=build and boot on volteer

Change-Id: Id03fe44ff859736f1e2f86231ca1e8d9dc4c50af
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2575198
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>