ec3po: console log printing in py2 and py3
The console log printing needs to work in py2 and py3. For this to work,
the 'data' passed in needs to be treated equally in both cases. Since
mostly what we care about is comparing it to other elements, we simply
ensure that the data becomes a list of integers in both py2 and py3, and
then we compare directly against the integer values of the characters we
want to handle in a special way (\b, \t, etc).

BUG=b:173654272
BRANCH=None
// this shows all the correctly parsed log output from the MCUs (ec, v4,
etc) in the logs
TEST=sudo servod -b scarlet
     cat /var/log/servod_9999/latest.DEBUG

Change-Id: Ibcbb95fd60619dc8234d7dd1f3d5d5c623542ce0
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>