chgstv2: OCPC: Don't always touch primary charger
On OCPC systems, when charging from an auxiliary charge port, we
should not unconditionally touch the primary charger as the charger
drivers may have configured the primary charger in a specific way
depending on the charging phase occurring.

This commit simply has the charger task not always touch the primary
charger when a charge_request comes in on OCPC systems.

BUG=b:174167890
BRANCH=dedede
TEST=Build and flash drawcia, discharge battery until DFET is
disabled, plug in 45W charger on sub-board, verify that precharge
currents are applied and when the battery requests a fast charge
current, the current is maintained at the precharge current level for
a few seconds.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I9d74bd8369ca0c5f6434ac5ec2c22bbbb2c11560
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2575838
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>