ec3po: handle passed commands from queue
The item coming out of the queue is a string in both py2 and py3. We
want to make sure that it's treated as a bytearray essentially so that
our comparisons work as expected. To achieve this, we simply call
encode() on the item before continuing to process it.

BUG=b:173654272
BRANCH=None
TEST=sudo servod -b scarlet
// servod startup turns off the timestamps for CPU uarts, exhibiting
this path

Change-Id: I7954930b658afb5b41fe8915144005d2552f3ad8
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>